refactor: extract authenticated user resolver
