+++ src/main/java/com/financebot/recurring/service/RecurringTransactionService.java
@@ -13,12 +13,12 @@ import com.financebot.recurring.mapper.RecurringTransactionMapper;
 import com.financebot.recurring.repository.RecurringTransactionRepository;
 import com.financebot.transaction.domain.TransactionType;
 import com.financebot.user.domain.User;
-import com.financebot.user.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import com.financebot.user.service.AuthenticatedUserResolver;
 
 import java.util.List;
 
@@ -27,17 +27,17 @@ import java.util.List;
 public class RecurringTransactionService {
 
     private final RecurringTransactionRepository recurringTransactionRepository;
-    private final UserRepository userRepository;
     private final AccountRepository accountRepository;
     private final CategoryRepository categoryRepository;
     private final RecurringTransactionMapper recurringTransactionMapper;
+    private final AuthenticatedUserResolver authenticatedUserResolver;
 
     @Transactional
     public RecurringTransactionResponse create(
             CreateRecurringTransactionRequest request,
             Authentication authentication
     ) {
-        User user = getAuthenticatedUser(authentication);
+        User user = authenticatedUserResolver.resolve(authentication);
         Account account = getUserAccount(request.accountId(), user.getId());
         Category category = getUserCategory(request.categoryId(), user.getId());
 
@@ -65,7 +65,7 @@ public class RecurringTransactionService {
 
     @Transactional(readOnly = true)
     public List<RecurringTransactionResponse> findAll(Authentication authentication) {
-        User user = getAuthenticatedUser(authentication);
+        User user = authenticatedUserResolver.resolve(authentication);
 
         return recurringTransactionRepository.findAllByUserIdOrderByCreatedAtDesc(user.getId())
                 .stream()
@@ -75,7 +75,7 @@ public class RecurringTransactionService {
 
     @Transactional(readOnly = true)
     public RecurringTransactionResponse findById(Long id, Authentication authentication) {
-        User user = getAuthenticatedUser(authentication);
+        User user = authenticatedUserResolver.resolve(authentication);
         RecurringTransaction recurringTransaction = getUserRecurringTransaction(id, user.getId());
 
         return recurringTransactionMapper.toResponse(recurringTransaction);
@@ -87,7 +87,7 @@ public class RecurringTransactionService {
             UpdateRecurringTransactionRequest request,
             Authentication authentication
     ) {
-        User user = getAuthenticatedUser(authentication);
+        User user = authenticatedUserResolver.resolve(authentication);
         RecurringTransaction recurringTransaction = getUserRecurringTransaction(id, user.getId());
 
         Account account = getUserAccount(request.accountId(), user.getId());
@@ -119,7 +119,7 @@ public class RecurringTransactionService {
 
     @Transactional
     public void delete(Long id, Authentication authentication) {
-        User user = getAuthenticatedUser(authentication);
+        User user = authenticatedUserResolver.resolve(authentication);
         RecurringTransaction recurringTransaction = getUserRecurringTransaction(id, user.getId());
 
         recurringTransactionRepository.delete(recurringTransaction);
@@ -127,7 +127,7 @@ public class RecurringTransactionService {
 
     @Transactional
     public RecurringTransactionResponse activate(Long id, Authentication authentication) {
-        User user = getAuthenticatedUser(authentication);
+        User user = authenticatedUserResolver.resolve(authentication);
         RecurringTransaction recurringTransaction = getUserRecurringTransaction(id, user.getId());
 
         recurringTransaction.setActive(true);
@@ -142,7 +142,7 @@ public class RecurringTransactionService {
 
     @Transactional
     public RecurringTransactionResponse deactivate(Long id, Authentication authentication) {
-        User user = getAuthenticatedUser(authentication);
+        User user = authenticatedUserResolver.resolve(authentication);
         RecurringTransaction recurringTransaction = getUserRecurringTransaction(id, user.getId());
 
         recurringTransaction.setActive(false);
@@ -164,15 +164,6 @@ public class RecurringTransactionService {
     private Category getUserCategory(Long categoryId, Long userId) {
         return categoryRepository.findByIdAndUserId(categoryId, userId)
                 .orElseThrow(() -> new EntityNotFoundException("Category not found"));
-    }
-
-    private User getAuthenticatedUser(Authentication authentication) {
-        if (authentication == null || authentication.getName() == null || authentication.getName().isBlank()) {
-            throw new IllegalArgumentException("Authenticated user is invalid");
-        }
-
-        return userRepository.findByEmail(authentication.getName())
-                .orElseThrow(() -> new EntityNotFoundException("Authenticated user not found"));
     }
 
     private void validateCategoryMatchesTransactionType(Category category, TransactionType transactionType) {
